refactor: don't remove fmt: on or fmt: off tags
I want these to be kept, so that I can use them in sleek without them
disappearing after the first time. This makes sure that running the
formatter twice gives the same results. Without this change, the first
run removes the format tags while not formatting the excluded section,
but then the next run (e.g. on save) formats everything as if the tags
were never there in the first place